Update intid to v0.3.0-beta.1, fixing breakage
The main issues is that the ContiguousIntegerId trait was renamed to IntegerIdContiguous.
There should be no more severe breakage now that we are in the beta channel.

Also taking advantage of the new `define_newtype_counter!` should help
insulate from the changes if they occur in the future.

Bump version to v0.1.4